etl/requirements.txt
pandas>=2.0
mysql-connector-python>=8.0